Add flags compílation

+++ cpp00/ex01/includes/PhoneBook.hpp
@@ -6,7 +6,7 @@
 /*   By: rlins <rlins@student.42sp.org.br>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/04/24 11:01:55 by rlins             #+#    #+#             */
-/*   Updated: 2023/04/25 16:32:40 by rlins            ###   ########.fr       */
+/*   Updated: 2023/04/25 16:41:16 by rlins            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,8 +26,8 @@ class PhoneBook
 		void	displaySingleContact(std::string input);
 		void	truncate(std::string str);
 		void	headerContacts(void);
-		bool	validInput(std::string input, const std::string allowedChar, int minLength);
-		bool	isValidField(std::string input, bool justNumber); // TODO: Não consigo entender o
+		bool	validInput(std::string input, const std::string allowedChar, long unsigned int minLength);
+		bool	isValidField(std::string input, bool justNumber);
 
 
 	public:
@@ -35,7 +35,6 @@ class PhoneBook
 		PhoneBook();
 		~PhoneBook();
 
-		/*TODO: Tem q ser publico mesmo?*/
 		void	menu();
 		void	farewell();
